refactor(nf): address PR #2904 review feedback
- Merge SynapseCrossTabBarPlot into SynapsePlot via crossTabConfig option;
  delete standalone component and update OrganizationDetailsTab to use SynapsePlot directly
- Replace MarkdownSynapse preview widget with EntityPreview in OrganizationAnalyticsTab
- Delete unused NFTopBar (redundant with navbar searchPath)
- Extract NFStatsBanner inline styles to NFStatsBanner.scss
