+++ apps/portals/nf/src/components/NFStatsBanner.scss
@@ -1,0 +1,63 @@
+.nf-stats-banner {
+  background: #fff;
+  border-bottom: 1px solid #dde3ea;
+
+  &__grid {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    padding: 24px 0;
+    gap: 24px;
+
+    @media (min-width: 600px) {
+      grid-template-columns: repeat(3, 1fr);
+    }
+
+    @media (min-width: 900px) {
+      grid-template-columns: repeat(5, 1fr);
+      gap: 0;
+      padding: 32px 0;
+    }
+  }
+
+  &__stat {
+    display: flex;
+    flex-direction: column;
+    gap: 4px;
+    text-decoration: none;
+    transition: opacity 0.15s;
+
+    @media (min-width: 900px) {
+      padding: 0 24px;
+      border-right: 1px solid #dde3ea;
+
+      &:last-child {
+        border-right: none;
+      }
+    }
+
+    &:hover {
+      opacity: 0.7;
+    }
+  }
+
+  &__number {
+    font-size: 2.25rem;
+    font-weight: 900;
+    line-height: 1;
+    color: #1b2a41;
+    letter-spacing: -0.02em;
+    font-variant-numeric: tabular-nums;
+
+    @media (min-width: 900px) {
+      font-size: 2.5rem;
+    }
+  }
+
+  &__label {
+    font-size: 0.6875rem;
+    font-weight: 700;
+    text-transform: uppercase;
+    letter-spacing: 0.14em;
+    color: #6b7f94;
+  }
+}
