+++ apps/portals/nf/src/components/NFStatsBanner.tsx
@@ -1,4 +1,4 @@
-import { Box, Container, Skeleton, Typography } from '@mui/material'
+import { Container, Skeleton } from '@mui/material'
 import { Link } from 'react-router'
 import useGetQueryResultBundle from 'synapse-react-client/synapse-queries/entity/useGetQueryResultBundle'
 import { parseEntityIdFromSqlStatement } from 'synapse-react-client/utils/functions/SqlFunctions'
@@ -10,6 +10,7 @@ import {
   studiesSql,
   toolsSql,
 } from '../config/resources'
+import './NFStatsBanner.scss'
 
 function StatCount({ sql }: { sql: string }) {
   const entityId = parseEntityIdFromSqlStatement(sql)
@@ -45,71 +46,19 @@ const STATS = [
 
 export default function NFStatsBanner() {
   return (
-    <Box
-      sx={{
-        background: '#fff',
-        borderBottom: '1px solid #DDE3EA',
-      }}
-    >
+    <div className="nf-stats-banner">
       <Container maxWidth="lg">
-        <Box
-          sx={{
-            display: 'grid',
-            gridTemplateColumns: {
-              xs: 'repeat(2, 1fr)',
-              sm: 'repeat(3, 1fr)',
-              md: 'repeat(5, 1fr)',
-            },
-            py: { xs: 3, md: 4 },
-            gap: { xs: 3, md: 0 },
-          }}
-        >
-          {STATS.map(({ label, sql, href }, i) => (
-            <Box
-              key={label}
-              component={Link}
-              to={href}
-              sx={{
-                textDecoration: 'none',
-                display: 'flex',
-                flexDirection: 'column',
-                gap: 0.5,
-                px: { md: 3 },
-                borderRight: {
-                  md: i < STATS.length - 1 ? '1px solid #DDE3EA' : 'none',
-                },
-                transition: 'opacity 0.15s',
-                '&:hover': { opacity: 0.7 },
-              }}
-            >
-              <Typography
-                component="div"
-                sx={{
-                  fontSize: { xs: '2.25rem', md: '2.5rem' },
-                  fontWeight: 900,
-                  lineHeight: 1,
-                  color: '#1B2A41',
-                  letterSpacing: '-0.02em',
-                  fontVariantNumeric: 'tabular-nums',
-                }}
-              >
+        <div className="nf-stats-banner__grid">
+          {STATS.map(({ label, sql, href }) => (
+            <Link key={label} to={href} className="nf-stats-banner__stat">
+              <div className="nf-stats-banner__number">
                 <StatCount sql={sql} />
-              </Typography>
-              <Typography
-                sx={{
-                  fontSize: '0.6875rem',
-                  fontWeight: 700,
-                  textTransform: 'uppercase',
-                  letterSpacing: '0.14em',
-                  color: '#6B7F94',
-                }}
-              >
-                {label}
-              </Typography>
-            </Box>
+              </div>
+              <div className="nf-stats-banner__label">{label}</div>
+            </Link>
           ))}
-        </Box>
+        </div>
       </Container>
-    </Box>
+    </div>
   )
 }
