+++ apps/portals/nf/src/pages/OrganizationDetailsPage/OrganizationAnalyticsTab.tsx
@@ -1,7 +1,7 @@
 import { DetailsPageContent } from '@sage-bionetworks/synapse-portal-framework/components/DetailsPage/DetailsPageContentLayout'
 import { useDetailsPageContext } from '@sage-bionetworks/synapse-portal-framework/components/DetailsPage/DetailsPageContext'
 import { Box, Chip, Paper, Typography } from '@mui/material'
-import MarkdownSynapse from 'synapse-react-client/components/Markdown/MarkdownSynapse'
+import EntityPreview from 'synapse-react-client/components/FilePreview/EntityPreview'
 
 // Synapse entity IDs of per-funder HTML dashboards generated by build_funder_dashboard.py.
 // Update after running: python build_funder_dashboard.py --funder <FUNDER> --upload
@@ -225,9 +225,7 @@ function OrganizationAnalyticsTab() {
           id: 'Downloads by Assay Type',
           title: 'Downloads by Assay Type',
           element: dashboardEntityId ? (
-            <MarkdownSynapse
-              markdown={`\${preview?entityId=${dashboardEntityId}}`}
-            />
+            <EntityPreview entityId={dashboardEntityId} />
           ) : (
             <SnowflakePlaceholder
               title="Representative assay types downloaded (Jan–Jun 2026)"
@@ -295,11 +293,7 @@ function OrganizationAnalyticsTab() {
               {
                 id: 'Full Analytics Dashboard',
                 title: 'Full Analytics Dashboard',
-                element: (
-                  <MarkdownSynapse
-                    markdown={`\${preview?entityId=${dashboardEntityId}}`}
-                  />
-                ),
+                element: <EntityPreview entityId={dashboardEntityId} />,
               },
             ]
           : []),
